requirements.txt
fastapi
uvicorn
httpx
mysql-connector-python
pytest
python-dotenv
pytest-dotenv